pyproject.toml
[tool.poetry]
name = "open-data-spain"
version = "0.1.0"
description = "API de servicios públicos Españoles"
authors = ["Alice & Bob LLC"]
license = "Todos los derechos reservados"
readme = "README.md"
packages = [{include = "open_data_spain"}]

[tool.poetry.dependencies]
python = "^3.11"
django = {extras = ["argon2"], version = "^5.0"}
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
dj-database-url = "^2.0.0"
django-model-utils = "^4.3.1"
unidecode = "^1.3.6"
celery = {extras = ["redis"], version = "^5.3.1"}
requests = "^2.31.0"
lxml = "^4.9.2"
html2text = "^2020.1.16"
dictionary-search = "^1.0.0"
googlemaps = "^4.10.0"
boto3 = "^1.34.3"
orjson = "^3.9.1"
drf-yasg = {extras = ["validation"], version = "^1.21.6"}
gunicorn = "^21.2.0"
gevent = "^23.9.1"
minio = "^7.1.15"
flower = "^2.0.0"
django-json-widget = "^1.1.1"
hiredis = "^2.3.2"
msgpack = "^1.0.5"
psycopg = {extras = ["binary"], version = "^3.1.15"}
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
sentry-sdk = "^1.39.1"
drf-spectacular = "^0.27.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
mongoengine = "^0.27.0"
django-allauth = "^0.59"
crispy-bootstrap5 = "^2023.10"
django-prettyjson = "^0.4.1"
dj-stripe = "^2.8.3"
jsonschema-rs = "^0.17.1"
motor = {extras = ["gssapi", "ocsp", "srv"], version = "^3.3.2"}
opencage = "^2.4.0"


[tool.poetry.group.dev.dependencies]
pyopenssl = "^23.3.0"
django-extensions = "^3.2.3"
werkzeug = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
